feat: Vault leveraging aave on eth and pancake on base part 1 (#340)

+++ packages/encoder-utils/src/libraries/forwarder/solidity_types.rs
@@ -3,7 +3,7 @@ use alloy_sol_types::sol;
 sol! {
     struct ForwardingConfig {
         address tokenAddress;
-        uint128 maxAmount;
+        uint256 maxAmount;
     }
 
     enum IntervalType {
